versão final do projeto
por mais que ja faça quase um ano que finalizei este projeto nunca tinha postado então resolvi postar agora

+++ delete_game.php
@@ -4,7 +4,6 @@ include('db_connect.php');
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
 
-    // Apaga a imagem do servidor
     $result = $conn->query("SELECT image_filename FROM jogos WHERE id = $id");
     $game = $result->fetch_assoc();
     $image_path = 'uploads/' . $game['image_filename'];
@@ -12,10 +11,20 @@ if (isset($_GET['id'])) {
         unlink($image_path);
     }
 
-    // Apaga o registro do banco de dados
+   
+    $stmt = $conn->prepare("DELETE FROM cart WHERE game_id = ?");
+    $stmt->bind_param("i", $id);
+    if (!$stmt->execute()) {
+        echo "Erro ao excluir registros dependentes: " . $stmt->error;
+        $stmt->close();
+        $conn->close();
+        exit();
+    }
+    $stmt->close();
+
+   
     $stmt = $conn->prepare("DELETE FROM jogos WHERE id = ?");
     $stmt->bind_param("i", $id);
-
     if ($stmt->execute()) {
         header("Location: manage_games.php");
     } else {
